% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{bootstrap_nongaussian}
\alias{bootstrap_nongaussian}
\title{Bootstrapping for non-gaussian functions (internal use)}
\usage{
bootstrap_nongaussian(
  bootstr,
  R_pe,
  formula,
  data,
  Ysim,
  mod,
  grname,
  grname_org,
  nboot,
  parallel,
  ncores,
  CI,
  rptObj,
  update
)
}
\arguments{
\item{bootstr}{bootstrap function. Re-assigns response simulated by simulate.merMod to data and estimates R with the R_pe function.}

\item{R_pe}{Function to estimate Repeatabilities and Variances for grouping factors, Residuals, Overdispersion and Fixed effects.}

\item{formula}{lme4 model formula}

\item{data}{data.frame given as original input}

\item{Ysim}{data.frame with simulated response variables from simulate.merMod}

\item{mod}{fitted lme4 model}

\item{grname}{original grnames vector without Residual or Fixed}

\item{grname_org}{original grnames vector}

\item{nboot}{number of bootstraps, equal to columns in Ysim}

\item{parallel}{boolean}

\item{ncores}{number of cores specified, defaults to NULL}

\item{CI}{confidence interval, defaults to 0.95}
}
\description{
Bootstrapping for non-gaussian functions (internal use)
}
\keyword{internal}
