% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_inputs}
\alias{check_inputs}
\title{Checks for data argument and uses non-standard evaluation}
\usage{
check_inputs(y, groups, data = NULL)
}
\arguments{
\item{y}{response variable}

\item{groups}{grouping variable}

\item{data}{data.frame containing variables}
}
\value{
returns a list containing the response and the grouping variable
}
\description{
NOT in USE at the moment
}

