% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{LRT_nongaussian}
\alias{LRT_nongaussian}
\title{Likelihood ratio test for non-gaussian functions (internal use)}
\usage{
LRT_nongaussian(formula, data, grname, mod, link, family)
}
\arguments{
\item{formula}{lme4 model formula}

\item{data}{data.frame given as original input}

\item{grname}{original grnames vector without Residual or Fixed}

\item{mod}{fitted lme4 model}

\item{link}{link function}

\item{family}{respnse family (so far just binomial or poisson)}
}
\description{
Likelihood ratio test for non-gaussian functions (internal use)
}
\keyword{internal}

