% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{overdisp}
\alias{overdisp}
\title{Calculate overdispersion}
\usage{
overdisp(model)
}
\arguments{
\item{model}{merMod object}
}
\value{
dispersion parameter
}
\description{
Calculate overdispersion
}
\details{
From http://glmm.wikidot.com/faq
}

