% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.rpt}
\alias{print.rpt}
\title{Print a rpt object}
\usage{
print.rpt(x, ...)
}
\arguments{
\item{x}{An rpt object returned from one of the rpt functions}

\item{\dots}{Additional arguments; none are used in this method.}
}
\description{
Displays the results a rpt object (i.e. the result of a rpt function call) an a nice form.
}
\examples{
# repeatability estimation for weight (body mass)
data(BodySize)
attach(BodySize)
print.rpt(rpt.Weight <- rpt.mcmcLMM(Weight, BirdID))
print(rpt.Weight)  # alternative call to print.rpt() through pring()
detach(BodySize)
}
\author{
Holger Schielzeth  (holger.schielzeth@ebc.uu.se) &
     Shinichi Nakagawa (shinichi.nakagawa@otago.ac.nz)
}
\references{
Nakagawa, S. and Schielzeth, H. (2010) \emph{Repeatability for Gaussian and
             non-Gaussian data: a practical guide for biologists}. Biological Reviews 85: 935-956
}
\seealso{
\link{rpt}, \link{rpt.corr}, \link{rpt.aov}, \link{rpt.remlLMM}, \link{rpt.mcmcLMM},
         \link{rpt.binomGLMM.add}, \link{rpt.binomGLMM.multi}, \link{rpt.poisGLMM.add}, \link{rpt.poisGLMM.multi}
}
\keyword{models}

