% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.rpt.R
\name{print.summary.rpt}
\alias{print.summary.rpt}
\title{Prints the summary of a rpt object}
\usage{
\method{print}{summary.rpt}(x, ...)
}
\arguments{
\item{x}{An rpt object returned from one of the rpt functions}

\item{\dots}{Additional arguments; none are used in this method.}
}
\description{
Displays the summary of an rpt object (i.e. the result of a rpt function call) in an 
extended form.
}
\author{
Holger Schielzeth  (holger.schielzeth@uni-jena.de), 
        Shinichi Nakagawa (s.nakagawa@unsw.edu.au),
        Martin Stoffel (martin.adam.stoffel@gmail.com)
}
\references{
Nakagawa, S. and Schielzeth, H. (2010) \emph{Repeatability for Gaussian and 
non-Gaussian data: a practical guide for biologists}. Biological Reviews 85: 935-956
}
\keyword{models}

