% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpt.binomGLMM.add.R
\name{rpt.binomGLMM.add}
\alias{rpt.binomGLMM.add}
\alias{rpt.binomGLMM.add_}
\title{GLMM-based Repeatability Using MCMC for Binomial Data}
\usage{
rpt.binomGLMM.add(data = NULL, y, groups, CI = 0.95, prior = NULL,
  verbose = FALSE, ...)

rpt.binomGLMM.add_(data = NULL, y, groups, CI = 0.95, prior = NULL,
  verbose = FALSE, ...)
}
\arguments{
\item{data}{\code{data.frame} containing response and groups variables.}

\item{y}{If binary, name of response variable in the \code{data.frame}. For proportion data,
specify two variables:
The first is the variable for number of successes \emph{m} and the second
variable is \emph{n-m}, where
\emph{m} is the number of successes and \emph{n} is the number of trials.}

\item{groups}{Name of group variable in the \code{data.frame}.}

\item{CI}{Width of the Bayesian credible interval (defaults to 0.95)}

\item{prior}{List of prior values passed to the \link{MCMCglmm} function
in \pkg{MCMCglmm} (see there for more details). Default priors will be
used if prior is \code{NULL}.}

\item{verbose}{Whether or not \link{MCMCglmm} should print MH diagnostics
are printed to screen. Defaults to FALSE.}

\item{...}{Additonal arguments that are passed on to \link{MCMCglmm}
      (e.g. length of chain, thinning interval).}
}
\value{
Returns an object of class rpt that is a a list with the following elements:
\item{datatype}{Type of response (here: 'binomial').}
\item{method}{Method used to calculate repeatability (here: 'MCMC').}
\item{CI}{Width of the Bayesian credibility interval.}
\item{R.link}{Point estimate for repeatability on the link scale, i.e. the mode of the posterior distribution.}
\item{se.link}{Standard error (\emph{se}) for the repeatability on the link scale, i.e. the standard deviation of the posterior distribution. Note that the distribution might not be symmetrical, in which case \emph{se} is less informative.}
\item{CI.link}{Bayesian credibility interval for the intraclass correlation (or repeatability) on the link scale based on the posterior distribution of \emph{R}.}
\item{P.link}{Significance test for the link scale repeatability, returned as \code{NA}, since the Bayesian approach conflicts with the null hypothesis testing.}
\item{R.org}{Point estimate for repeatability on the original scale, i.e. the mode of the posterior distribution.}
\item{se.org}{Standard error (\emph{se}) for repeatability on the original scale, i.e. the standard deviation of the posterior distribution. Note that the distribution might not be symmetrical, in which case \emph{se} is less informative.}
\item{CI.org}{Bayesian credibility interval for repeatability on the original scale based on the posterior distribution of \emph{R}.}
\item{P.org}{Significance test for the original scale repeatability, returned as \emph{NA}, since the Bayesian approach conflicts with the null hypothesis testing.}
\item{R.post}{Named list of MCMC samples form the posterior distributions. \code{R.link} gives the samples for the link scale repeatability, \code{R.org} gives the samples for the original scale repeatability.}
\item{MCMCpars}{Burnin, length of chain, thinning interval of MCMC chain.}
\item{ngroups}{Number of groups.}
\item{nobs}{Number of observations.}
\item{mod}{Fitted model.}
}
\description{
Calculates repeatability from a generalised linear mixed-effects models
fitted by MCMC for binary and proportion data.
}
\details{
Models are fitted using the \link{MCMCglmm} function in \pkg{MCMCglmm}.
     The categorical family is used for binary data, while the multinomial2 is used for proportion data.
     Models for binary data are fitted with \code{list(R=list(V=1,fix=1),G=list(G1=list(V=1,nu=1,alpha.mu=0,alpha.V=25^2)))} unless other priors are specified in the call.
     Models for proportion data are fitted with \code{list(R=list(V=1e-10,nu=-1),G=list(G1=list(V=1,nu=1,alpha.mu=0,alpha.V=25^2)))} unless other priors are specified in the call.
}
\examples{
# repeatability estimations for egg dumping (binary data)
     data(BroodParasitism)
     EggDump <- subset(BroodParasitism, OwnClutchesBothSeasons == 1,
                       select = c(HostYN, FemaleID))
     (rpt.Host <- rpt.binomGLMM.add(data = EggDump, HostYN, FemaleID))
     (rpt.BroodPar <- rpt.binomGLMM.add(data = BroodParasitism, cbpYN, FemaleID))

# repeatability estimations for egg dumping (proportion data)
     data(BroodParasitism)
     ParasitismOR <- subset(BroodParasitism, OwnClutchesBothSeasons == 1,
                            select= c(HostClutches, OwnClutches, FemaleID))
     ParasitismOR$parasitised <-  ParasitismOR$OwnClutches -
                                  ParasitismOR$HostClutches
     (rpt.Host <- rpt.binomGLMM.add(data = ParasitismOR,
                                    y = list(HostClutches, parasitised),
                                    groups = FemaleID))
}
\author{
Holger Schielzeth  (holger.schielzeth@ebc.uu.se) &
        Shinichi Nakagawa (shinichi.nakagawa@otago.ac.nz) &
        Martin Stoffel (martin.adam.stoffel@gmail.com)
}
\references{
Browne, W. J., Subramanian, S. V., et al. (2005). \emph{Variance partitioning in multilevel logistic models that exhibit overdispersion}. Journal of the Royal Statistical Society A 168: 599-613. \cr

Goldstein, H., Browne, W., et al. (2002). \emph{Partitioning variation in multilevel models}. Understanding Statistics 1: 223-231. \cr

Nakagawa, S. and Schielzeth, H. (2010) \emph{Repeatability for Gaussian and non-Gaussian data: a practical guide for biologists}. Biological Reviews 85: 935-956
}
\seealso{
\link{rpt.binomGLMM.multi}, \link{rpt}, \link{summary.rpt}, \link{plot.rpt}
}
\keyword{models}

