% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rpt.binomGLMM.multi.adj}
\alias{rpt.binomGLMM.multi.adj}
\title{GLMM-based Repeatability Using PQL Estimation for Binomial Data}
\usage{
rpt.binomGLMM.multi.adj(formula, grnames, data, link = c("logit", "probit"),
  CI = 0.95, nboot = 1000, npermut = 1000, parallel = FALSE,
  ncores = 0)
}
\arguments{
\item{formula}{Formula as used e.g. by \link{glmmPQL}. The grouping factor of
interest needs to be included as a random effect, e.g. '(1|groups)'.
Covariates and additional random effects can be included to estimate adjusted repeatabilities.}

\item{data}{Data frame containing respnse and groups variable.}

\item{link}{Link function, \code{log} and \code{sqrt} are allowed, defaults
to \code{log}.}

\item{CI}{Width of the confidence interval (defaults to 0.95).}

\item{nboot}{Number of parametric bootstraps for interval estimation
(defaults to 1000). Larger numbers of permutations give a better
asymtotic CI, but may be very time-consuming.}

\item{npermut}{Number of permutations for significance testing
(defaults to 1000). Larger numbers of permutations give better
asymtotic \emph{P} values, but may be very time-consuming.}

\item{parallel}{If TRUE, bootstraps will be distributed.}

\item{ncores}{Specify number of cores to use for parallelization. On default,
       all cores are used.}

\item{grname}{A character string or vector of character strings giving the
name(s) of the grouping factor(s), for which the repeatability should
be estimated. Spelling needs to match the random effect names as given in \code{formula}.}
}
\value{
Returns an object of class rpt that is a a list with the following elements:
\item{datatype}{Type of response (here: "binomial").}
\item{method}{Method used to calculate repeatability (here: "PQL").}
\item{link}{Link function used (here: "logit" or "probit").}
\item{CI}{Width of the confidence interval.}
\item{R.link}{Point estimate for repeatability on the link scale.}
\item{se.link}{Standard error  (\emph{se}) for repeatability on the link
      scale, i.e. the standard deviation of the parametric bootstrap runs.
      Note that the distribution might not be symmetrical, in which case
      the \emph{se} is less informative.}
\item{CI.link}{Confidence interval for repeatability on the link scale based
      on parametric-boostrapping of \emph{R}.}
\item{P.link}{Approximate \emph{P} value from a significance test for the
      link scale repeatability based on randomisation.}
\item{R.org}{Point estimate for repeatability \emph{R} on the original scale.}
\item{se.org}{Standard error  (\emph{se}) for repeatability on the original
      scale, i.e. the standard deviation of the parametric bootstrap runs.
      Note that the distribution might not be symmetrical, in which case
       \emph{se} is less informative.}
\item{CI.org}{Confidence interval for repeatability on the link scale based
      on parametric-boostrapping of \emph{R}.}
\item{P.org}{Approximate \emph{P} value from a a significance test for the
      original scale repeatability based on randomisation. }
\item{omega}{Multiplicative overdispersion parameter.}
\item{R.boot}{Named list of parametric bootstap samples for \emph{R}.
      \code{R.link} gives the samples for the link scale repeatability,
      \code{R.org} gives the samples for the original scale repeatability.}
\item{R.permut}{Named list of permutation samples for \emph{R}. \code{R.link}
      gives the samples for the link scale repeatability, \code{R.org} gives
      the samples for the original scale repeatability.}
\item{ngroups}{Number of groups.}
\item{nobs}{Number of observations.}
}
\description{
Calculates repeatability from a generalised linear mixed-effects models fitted
by PQL (penalized-quasi likelihood) estimation for binary and proportion data.
}
\details{
Models are fitted using the \link{glmmPQL} function in \pkg{MASS}
         with the \code{quasibinomial} family (proportion data) or the
         \code{binomial} family (binary data). }\note{ Confidence intervals
         and standard errors are inappropriate at high repeatabilities
         (\emph{omega} < 1), because parametric bootstrapping allows only
          \emph{omega} greater than or equal to 1.
}
\examples{
# repeatability estimations for egg dumping (binary data)
     data(BroodParasitism)
     EggDump <- subset(BroodParasitism, OwnClutchesBothSeasons == 1, select = c(HostYN, FemaleID))
     (rpt.Host <- rpt.binomGLMM.multi("HostYN", "FemaleID", data = EggDump,  nboot=10, npermut=10))
     # low number of nboot and npermut to speed up error checking
     (rpt.BroodPar <- rpt.binomGLMM.multi("cbpYN", "FemaleID", data = BroodParasitism, nboot=10, npermut=10))
     # low number of nboot and npermut to speed up error checking

# repeatability estimations for egg dumping (proportion data)
     data(BroodParasitism)
     ParasitisedOR <- subset(BroodParasitism,  select= c(HostClutches, OwnClutches, FemaleID))
     ParasitisedOR$parasitised <-  ParasitisedOR$OwnClutches - ParasitisedOR$HostClutches
     (rpt.Host <- rpt.binomGLMM.multi(c("HostClutches", "parasitised"), "FemaleID",
                                     data = ParasitisedOR[BroodParasitism$OwnClutchesBothSeasons == 1, ],
                                     nboot=10, npermut=10))
                                      # reduced number of npermut iterations

     ParasitismOR <- subset(BroodParasitism,  select= c(cbpEggs, nEggs, FemaleID))
     ParasitismOR$parasitised <-  ParasitismOR$nEggs  - ParasitismOR$cbpEggs
     zz = which(ParasitismOR[,1]==0 & ParasitismOR[,2]==0) # some rows have entries 0,0 and need to be removed
     (rpt.BroodPar <- rpt.binomGLMM.multi(c("cbpEggs", "parasitised"), "FemaleID",
                                             data = ParasitismOR[-zz, ], nboot=10, npermut=10))
}
\author{
Holger Schielzeth  (holger.schielzeth@ebc.uu.se) &
     Shinichi Nakagawa (shinichi.nakagawa@otago.ac.nz)
}
\references{
Browne, W. J., Subramanian, S. V., et al. (2005). \emph{Variance partitioning in multilevel logistic models that exhibit overdispersion}. Journal of the Royal Statistical Society A 168: 599-613. \cr

Goldstein, H., Browne, W., et al. (2002). \emph{Partitioning variation in multilevel models}. Understanding Statistics 1: 223-231. \cr

Nakagawa, S. and Schielzeth, H. (2010) \emph{Repeatability for Gaussian and non-Gaussian data: a practical guide for biologists}. Biological Reviews 85: 935-956
}
\seealso{
\link{rpt.binomGLMM.add}, \link{rpt}, \link{print.rpt}
}
\keyword{#}
\keyword{models}

