% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpt.corr.R
\name{rpt.corr}
\alias{rpt.corr}
\title{Correlation-based Repeatability}
\usage{
rpt.corr(data = NULL, y, groups, CI = 0.95, nboot = 1000,
  npermut = 1000, parallel = FALSE, ncores = 0)
}
\arguments{
\item{data}{data.frame containing response and groups variable.}

\item{y}{Name of response variable in the data.frame. Missing values are not allowed.}

\item{groups}{Name of group variable in the data.frame (will be converted to a factor).
Note that each group identity has to appear exactly twice.}

\item{CI}{Width of the confidence interval between 0 and 1 (defaults to 0.95).}

\item{nboot}{Number of bootstrapping runs used when calculating an asymptotic
confidence interval (defaults to 1000).}

\item{npermut}{Number of permutations used when calculating asymptotic
\emph{P} values (defaults to 1000).}

\item{parallel}{If TRUE, bootstraps will be distributed.}

\item{ncores}{Specify number of cores to use for parallelization. On default,
       all cores are used.}
}
\value{
Returns an object of class rpt that is a a list with the following elements:
\item{datatype}{Response distribution (here: 'Gaussian').}
\item{method}{Method used to calculate repeatability (here: 'corr').}
\item{R}{Point estimate for repeatability \emph{R}.}
\item{se}{Asymptotic standard error for repeatability based on non-parametric bootstrapping.}
\item{CI}{Asymptotic confidence interval for repeatability based on non-parametric bootstrapping.}
\item{P}{Asymptotic \emph{P} value from a significance test for the intraclass correlation based on permutation.}
\item{R.permut}{Permutation samples for \emph{R}.}
\item{ngroups}{Number of groups.}
\item{nobs}{Number of observations.}
}
\description{
Calculates repeatability based on inter-class correlations
}
\examples{
# repeatability for male breeding success on a transformed scale
  data(Fledglings)
  Fledglings$sqrtFledge <- sqrt(Fledglings$Fledge)
  (rpt.Fledge <- rpt.corr(sqrtFledge, MaleID, data = Fledglings,
  nboot=10, npermut=10))  # reduced number of iterations
}
\author{
Holger Schielzeth  (holger.schielzeth@ebc.uu.se) &
        Shinichi Nakagawa (shinichi.nakagawa@otago.ac.nz) &
        Martin Stoffel (martin.adam.stoffel@gmail.com)
}
\references{
Sokal, R. R. and F. J. Rohlf (1995). \emph{Biometry: The principles and practice of statistics in biological research}. New York, W.H. Freeman and Company.

Nakagawa, S. and Schielzeth, H. (2010) \emph{Repeatability for Gaussian and non-Gaussian data: a practical guide for biologists}. Biological Reviews 85: 935-956
}
\seealso{
\link{rpt.aov}, \link{rpt.remlLMM}, \link{rpt.mcmcLMM}, \link{rpt}, \link{summary.rpt}, \link{plot.rpt}
}
\keyword{models}

