% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpt.remlLMM.R
\name{rpt.remlLMM}
\alias{rpt.remlLMM}
\alias{rpt.remlLMM_}
\title{LMM-based Repeatability Using REML}
\usage{
rpt.remlLMM(data = NULL, y, groups, CI = 0.95, nboot = 1000,
  npermut = 1000, parallel = FALSE, ncores = NULL)

rpt.remlLMM_(data = NULL, y, groups, CI = 0.95, nboot = 1000,
  npermut = 1000, parallel = FALSE, ncores = NULL)
}
\arguments{
\item{data}{data.frame containing response and groups variable.}

\item{y}{Name of response variable in the data.frame. Missing values are not allowed.}

\item{groups}{Name of group variable in the data.frame.}

\item{CI}{Width of the confidence interval (defaults to 0.95).}

\item{nboot}{Number of parametric bootstraps for interval estimation.
Defaults to 1000. Larger numbers of permutations give a better
asymtotic CI, but may be very time-consuming.}

\item{npermut}{Number of permutations used when calculating
asymptotic \emph{P} values (defaults to 1000).}

\item{parallel}{If TRUE, bootstraps will be distributed.}

\item{ncores}{Specify number of cores to use for parallelization. On default,
       all cores are used.}
}
\value{
Returns an object of class rpt that is a a list with the following elements:
\item{call}{Model call.}
\item{datatype}{Response distribution (here: 'Gaussian').}
\item{method}{Method used to calculate repeatability (here: 'REML').}
\item{CI}{Width of the confidence interval.}
\item{R}{Point estimate for repeatability.}
\item{se}{Approximate standard error (\emph{se}) for repeatability. Note that the distribution might not be symmetrical, in which case the \emph{se} is less informative.}
\item{CI.R}{Confidence interval for  repeatability.}
\item{P}{Vector of Approximate \emph{P} value from Likelihood-ratio test and Approximate \emph{P} value from a significance test based on permutation.}
\item{LRT}{Vector of Likelihood-ratios for the model and the reduced model, and \emph{P} value and degrees of freedom for the Likelihood-ratio test}
\item{R.boot}{Parametric bootstrap samples for \emph{R}.}
\item{R.permut}{Permutation samples for \emph{R}.}
\item{ngroups}{Number of groups.}
\item{nobs}{Number of observations.}
\item{mod}{Fitted model.}
}
\description{
Calculates repeatability from a linear mixed-effects models fitted by REML (restricted maximum likelihood).
}
\examples{
# repeatability estimation for tarsus length - a very high R
data(BodySize)
(rpt.BS <- rpt.remlLMM(Tarsus, BirdID, data = BodySize,
                       nboot=10, npermut=10))
# reduced number of nboot and npermut iterations

# repeatability estimation for weight (body mass) - a lower R
# than the previous one
data(BodySize)
(rpt.Weight <- rpt.remlLMM(Weight, BirdID, data = BodySize,
                            nboot=10, npermut=10))
# reduced number of nboot and npermut iterations
}
\author{
Holger Schielzeth  (holger.schielzeth@ebc.uu.se),
        Shinichi Nakagawa (shinichi.nakagawa@otago.ac.nz),
        Martin A. Stoffel (martin.adam.stoffel@gmail.com)
}
\references{
Carrasco, J. L. and Jover, L.  (2003). \emph{Estimating the generalized
concordance correlation coefficient through variance components}. Biometrics 59: 849-858.

Faraway, J. J. (2006). \emph{Extending the linear model with R}. Boca Raton, FL, Chapman & Hall/CRC.

Nakagawa, S. and Schielzeth, H. (2010) \emph{Repeatability for Gaussian and
             non-Gaussian data: a practical guide for biologists}. Biological Reviews 85: 935-956
}
\seealso{
\link{rpt.mcmcLMM}, \link{rpt.aov}, \link{rpt.corr}, \link{print.rpt}, \link{rpt.remlLMM.adj}
}
\keyword{models}

