% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptGaussianNew.R
\name{rptGaussianNew}
\alias{rptGaussianNew}
\title{GLMM-based Repeatability Using REML}
\usage{
rptGaussianNew(formula, grname, data, CI = 0.95, nboot = 1000,
  npermut = 1000, parallel = FALSE, ncores = NULL)
}
\arguments{
\item{formula}{Formula as used e.g. by \link{glmer}. The grouping factor(s) of
interest needs to be included as a random effect, e.g. '(1|groups)'.
Covariates and additional random effects can be included to estimate adjusted repeatabilities.}

\item{grname}{A character string or vector of character strings giving the
name(s) of the grouping factor(s), for which the repeatability should
be estimated. Spelling needs to match the random effect names as given in \code{formula}.}

\item{data}{A dataframe that contains the variables included in the \code{formula}
and \code{grname} arguments.}

\item{CI}{Width of the confidence interval (defaults to 0.95).}

\item{nboot}{Number of parametric bootstraps for interval estimation.
Defaults to 1000. Larger numbers of permutations give a better
asymtotic CI, but may be very time-consuming.}

\item{npermut}{Number of permutations used when calculating 
asymptotic \emph{P} values (defaults to 1000).}

\item{parallel}{If TRUE, bootstraps will be distributed.}

\item{ncores}{Specify number of cores to use for parallelization. On default,
all cores but one are used.}

\item{link}{Link function. \code{log} and \code{sqrt} are allowed, defaults to \code{log}.}
}
\description{
Calculates repeatability from a general linear mixed-effects models fitted by REML (restricted maximum likelihood).
}
\examples{
 
# repeatability estimation for tarsus length - a very high R
data(BodySize)
(rpt.BS <- rptGaussianNew(formula = Tarsus ~ 1 + (1|Sex) + (1|BirdID), grname = c('Sex', 'BirdID'), 
 data=BodySize, nboot=10, npermut=10))
 
 (rpt.BS <- rptGaussianNew(formula = Tarsus ~ 1 + (1|BirdID), grname = c('BirdID'), 
 data=BodySize, nboot=10, npermut=10))
}
\author{
Holger Schielzeth  (holger.schielzeth@ebc.uu.se),
        Shinichi Nakagawa (shinichi.nakagawa@otago.ac.nz) &
        Martin Stoffel (martin.adam.stoffel@gmail.com)
}
\references{
Carrasco, J. L. and Jover, L.  (2003). \emph{Estimating the generalized 
concordance correlation coefficient through variance components}. Biometrics 59: 849-858.

Faraway, J. J. (2006). \emph{Extending the linear model with R}. Boca Raton, FL, Chapman & Hall/CRC.

Nakagawa, S. and Schielzeth, H. (2010) \emph{Repeatability for Gaussian and 
non-Gaussian data: a practical guide for biologists}. Biological Reviews 85: 935-956
}
\seealso{
\link{rpt}
}

