% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptBernoulli.R
\name{rptProportion}
\alias{rptProportion}
\title{LMM-based Repeatability Using REML}
\usage{
rptProportion(formula, grname, data, link = c("logit", "probit"), CI = 0.95,
  nboot = 1000, npermut = 1000, parallel = FALSE, ncores = NULL)
}
\arguments{
\item{formula}{Formula as used e.g. by \link{glmer}. The grouping factor of
interest needs to be included as a random effect, e.g. '(1|groups)'.
Covariates and additional random effects can be included to estimate adjusted repeatabilities.}

\item{grname}{A character string or vector of character strings giving the
name(s) of the grouping factor(s), for which the repeatability should
be estimated. Spelling needs to match the random effect names as given in \code{formula}.}

\item{data}{A dataframe that contains the variables included in the formula argument.}

\item{CI}{Width of the confidence interval (defaults to 0.95).}

\item{nboot}{Number of parametric bootstraps for interval estimation.
Defaults to 1000. Larger numbers of permutations give a better
asymtotic CI, but may be very time-consuming.}

\item{npermut}{Number of permutations used when calculating 
asymptotic \emph{P} values (defaults to 1000). Currently not in use!}

\item{parallel}{If TRUE, bootstraps will be distributed.}

\item{ncores}{Specify number of cores to use for parallelization. On default,
all cores but one are used.}
}
\value{
Returns an object of class rpt that is a a list with the following elements: 
\item{datatype}{Response distribution (here: 'Gaussian').}
\item{method}{Method used to calculate repeatability (here: 'REML').}
\item{CI}{Width of the confidence interval.}
\item{R}{Point estimate for repeatability.}
\item{se}{Approximate standard error (\emph{se}) for repeatability. Note that the distribution might not be symmetrical, in which case the \emph{se} is less informative.}
\item{CI.R}{Confidence interval for  repeatability.}
\item{P}{Approximate \emph{P} value from a significance test based on likelihood-ratio.}
\item{P.permut}{Approximate \emph{P} value from a significance test based on permutation of residuals.}
\item{R.boot}{Parametric bootstrap samples for \emph{R}.}
\item{R.permut}{Permutation samples for \emph{R}.}
\item{LRT}{Vector of Likelihood-ratios for the model(s) and the reduced model(s), and \emph{P} value(s) and degrees of freedom for the Likelihood-ratio test} 
\item{ngroups}{Number of groups.}
\item{nobs}{Number of observations.}
\item{mod}{Fitted model.}
}
\description{
Calculates repeatability from a linear mixed-effects models fitted by REML (restricted maximum likelihood).
}
\examples{
 
# repeatability estimations for egg dumping (binary data)
data(BroodParasitism)
EggDump <- subset(BroodParasitism, OwnClutchesBothSeasons == 1, select = c(HostYN, FemaleID))
(rpt.Host <- rpt.binomGLMM.multi(data = EggDump, HostYN, FemaleID, nboot=10, npermut=10))  
# low number of nboot and npermut for speed reasons

(rpt.BroodPar <- rpt.binomGLMM.multi(data = BroodParasitism, cbpYN, FemaleID, 
                                     nboot=10, npermut=10))
  
  
# repeatability estimations for egg dumping (proportion data)
data(BroodParasitism)
ParasitisedOR <- subset(BroodParasitism,  select= c(HostClutches, OwnClutches, FemaleID))
ParasitisedOR$parasitised <-  ParasitisedOR$OwnClutches - ParasitisedOR$HostClutches 
(rpt.Host <- rpt.binomGLMM.multi(
                             data = ParasitisedOR[BroodParasitism$OwnClutchesBothSeasons == 1, ],
                             list(HostClutches, parasitised), groups = FemaleID, 
                             nboot=10, npermut=10))
                             
ParasitismOR <- subset(BroodParasitism,  select= c(cbpEggs, nEggs, FemaleID))
ParasitismOR$parasitised <-  ParasitismOR$nEggs  - ParasitismOR$cbpEggs
# some rows have entries 0,0 and need to be removed
zz = which(ParasitismOR[,1]==0 & ParasitismOR[,2]==0) 
(rpt.BroodPar <- rpt.binomGLMM.multi( data = ParasitismOR[-zz, ],
                                     y = list(cbpEggs, parasitised), groups = FemaleID,
                                     nboot = 10, npermut = 10))
rpt.BroodPar <- rptProportion(formula = cbind(cbpEggs, parasitised) ~ (1|FemaleID), data = ParasitismOR[-zz, ],
                             nboot = 10, npermut = 10)
}
\author{
Holger Schielzeth  (holger.schielzeth@ebc.uu.se),
        Shinichi Nakagawa (shinichi.nakagawa@otago.ac.nz) &
        Martin Stoffel (martin.adam.stoffel@gmail.com)
}
\references{
Carrasco, J. L. and Jover, L.  (2003). \emph{Estimating the generalized 
concordance correlation coefficient through variance components}. Biometrics 59: 849-858.

Faraway, J. J. (2006). \emph{Extending the linear model with R}. Boca Raton, FL, Chapman & Hall/CRC.

Nakagawa, S. and Schielzeth, H. (2010) \emph{Repeatability for Gaussian and 
non-Gaussian data: a practical guide for biologists}. Biological Reviews 85: 935-956
}
\seealso{
\link{rpt}
}

