% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptr-package.R
\docType{package}
\name{rptr-package}
\alias{rptr}
\alias{rptr-package}
\title{rptr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
