% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rpt.R
\name{summary.rpt}
\alias{summary.rpt}
\title{Summary of a rpt object}
\usage{
\method{summary}{rpt}(object, ...)
}
\arguments{
\item{object}{An rpt object returned from one of the rpt functions}

\item{\dots}{Additional arguments; none are used in this method.}
}
\description{
Summary of a rpt object
}
\author{
Holger Schielzeth  (holger.schielzeth@ebc.uu.se),
        Shinichi Nakagawa (shinichi.nakagawa@otago.ac.nz) &
        Martin Stoffel (martin.adam.stoffel@gmail.com)
}
\references{
Nakagawa, S. and Schielzeth, H. (2010) \emph{Repeatability for Gaussian and 
             non-Gaussian data: a practical guide for biologists}. Biological Reviews 85: 935-956
}
\keyword{models}

