% Generated by roxygen2 (4.0.2): do not edit by hand
\name{visualize.rpt}
\alias{visualize.rpt}
\title{Plot a rpt object}
\usage{
visualize.rpt(x)
}
\arguments{
\item{x}{An rpt object returned from one of the rpt functions}

\item{\dots}{Additional arguments; none are used in this method.}
}
\description{
Plots the distribution of repeatability estimates with a confidence interval
}
\examples{
# repeatability estimation for tarsus length - a very high R
data(BodySize)
attach(BodySize)
(rpt.BS <- rpt.remlLMM(Tarsus, BirdID, nboot=100, npermut=100))
# reduced number of nboot and npermut iterations
plot(rpt.BS)
detach(BodySize)
}
\author{
Holger Schielzeth  (holger.schielzeth@ebc.uu.se) &
     Shinichi Nakagawa (shinichi.nakagawa@otago.ac.nz)
}
\references{
Nakagawa, S. and Schielzeth, H. (2010) \emph{Repeatability for Gaussian and
             non-Gaussian data: a practical guide for biologists}. Biological Reviews 85: 935-956
}
\seealso{
\link{rpt}, \link{rpt.corr}, \link{rpt.aov}, \link{rpt.remlLMM}, \link{rpt.mcmcLMM},
         \link{rpt.binomGLMM.add}, \link{rpt.binomGLMM.multi}, \link{rpt.poisGLMM.add}, \link{rpt.poisGLMM.multi}
}
\keyword{models}

