\name{Fledglings}
\alias{Fledglings}
\docType{data}
\title{
	Male House Sparrow Fledgling Success Dataset
}
\description{
  Dataset containing fledgling success for male house sparrows per year over two years.
}
\usage{
	data(Fledglings)
}
\format{
  A data frame with 94 observations on the following 3 variables.
  \describe{
    \item{MaleID}{Male bird identities (47 birds in total).}
    \item{Age}{Age of the males (in years).}
    \item{Fledge}{Number of successfully fledged offspring.}
  }
}
\details{
	The number of fledgling male house sparrows had both in 2004 and in 2005 on Lundy Island, UK along with male age information.
}
\references{
	Nakagawa, S., Ockendon, N., et al. (2007)  \emph{Does the badge of status influence parental care and investment in house sparrows? An experimental test}. Oecologia 153: 749-760.
} 

\keyword{datasets}