% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rpt.R
\name{print.rpt}
\alias{print.rpt}
\title{Print a rpt object}
\usage{
\method{print}{rpt}(x, ...)
}
\arguments{
\item{x}{An rpt object returned from one of the rpt functions}

\item{\dots}{Additional arguments; none are used in this method.}
}
\description{
Displays the results a rpt object (i.e. the result of a rpt function call) in a nice form.
}
\author{
Holger Schielzeth  (holger.schielzeth@ebc.uu.se), 
        Shinichi Nakagawa (shinichi.nakagawa@otago.ac.nz) &
        Martin Stoffel (martin.adam.stoffel@gmail.com)
}
\references{
Nakagawa, S. and Schielzeth, H. (2010) \emph{Repeatability for Gaussian and 
             non-Gaussian data: a practical guide for biologists}. Biological Reviews 85: 935-956
}
\keyword{models}

