% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpt.aov.R
\name{rpt.aov}
\alias{rpt.aov}
\alias{rpt.aov_}
\title{ANOVA-based Repeatability}
\usage{
rpt.aov(data = NULL, y, groups, CI = 0.95, npermut = 1000,
  parallel = FALSE, ncores = NULL)

rpt.aov_(data = NULL, y, groups, CI = 0.95, npermut = 1000,
  parallel = FALSE, ncores = NULL)
}
\arguments{
\item{data}{data.frame containing response and groups variable.}

\item{y}{Name of the response variable in the data.frame}

\item{groups}{Name of the variable containing the group identities in the data.frame (will be converted to a factor).}

\item{CI}{Width of the confidence interval between 0 and 1 (defaults to 0.95).}

\item{npermut}{Number of permutations used when calculating asymptotic \emph{P} values (defaults to 1000).}

\item{parallel}{If TRUE, permutations will be distributed.}

\item{ncores}{Specify number of cores to use for parallelization. On default,
       all cores are used.}
}
\value{
Returns an object of class rpt that is a a list with the following elements:
\item{datatype}{Response distribution (here: 'Gaussian').}
\item{method}{Method used to calculate repeatability (here: 'ANOVA').}
\item{R}{Point estimate for the repeatability (denoted as \emph{R}).}
\item{se}{Asymptotic standard error for repeatability (ICC) based Becker (1982).}
\item{CI.R}{Asymptotic confidence interval for repeatability.}
\item{P}{Named vector of two \emph{P} values (significance tests): \code{P.aov} is the \emph{P} value for the ANOVA F test, \code{P.permut} is the permutation based \emph{P} value.}
\item{R.permut}{Repeatability \emph{R} estimates for each permutation run.}
\item{ngroups}{Number of groups.}
\item{nobs}{Number of observations.}
\item{mod}{Anova model.}
}
\description{
Repeatability calculations based on Analysis of Variance (ANOVA).
}
\examples{
# repeatability estimation for tarsus length - a very high R
# reduced number of npermut iterations
     data(BodySize)
     (rpt.BS <- rpt.aov(Tarsus, BirdID, data = BodySize, npermut=10))

# repeatability estimation for weight (body mass) - a lower R
# than the previous one
     data(BodySize)
     (rpt.Weight <- rpt.aov(data = BodySize, y = Weight, groups = BirdID, CI = 0.95, npermut=10))
}
\author{
Holger Schielzeth  (holger.schielzeth@ebc.uu.se) &
        Shinichi Nakagawa (shinichi.nakagawa@otago.ac.nz) &
        Martin Stoffel (martin.adam.stoffel@gmail.com)
}
\references{
Becker, W. A. (1992) \emph{A manual of quantitative genetics}. 5th edn. Academic Enterprises, Pullman, WA. \cr

Lessells, C. M. and Boag, P. T. (1987) \emph{Unrepeatable repeatabilities: a common mistake}. Auk 104: 116-121. \cr

Nakagawa, S. and Schielzeth, H. (2010) \emph{Repeatability for Gaussian and non-Gaussian data: a practical guide for biologists}. Biological Reviews 85: 935-956
}
\seealso{
\link{rpt.corr}, \link{rpt.remlLMM}, \link{rpt.mcmcLMM}, \link{rpt}, \link{summary.rpt}, \link{plot.rpt}
}
\keyword{models}

