% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpt.R
\name{rpt}
\alias{rpt}
\title{(Adjusted) Repeatability Calculation for Gaussian and Non-Gaussian Data}
\usage{
rpt(formula, grname, data, datatype = c("Gaussian", "Binomial", "Proportion",
  "count"), link = c("logit", "probit", "log", "sqrt"), CI = 0.95,
  nboot = 1000, npermut = 0, parallel = FALSE, ncores = NULL)
}
\arguments{
\item{formula}{Formula as used e.g. by \link{lmer}. The grouping factor(s) of
interest needs to be included as a random effect, e.g. '(1|groups)'.
Covariates and additional random effects can be included to estimate adjusted 
repeatabilities.}

\item{grname}{A character string or vector of character strings giving the
name(s) of the grouping factor(s), for which the repeatability should
be estimated. Spelling needs to match the random effect names as given in \code{formula} 
and terms have to be set in quotation marks.}

\item{data}{A dataframe that contains the variables included in the \code{formula}
and \code{grname} arguments.}

\item{datatype}{Character string specifying the data type ('Gaussian', 
'Binary', 'Proportion', 'Poisson').}

\item{link}{Character string specifying the link function. Ignored for 
'Gaussian' datatype.}

\item{CI}{Width of the required confidence interval between 0 and 1 (defaults to 
0.95).}

\item{nboot}{Number of parametric bootstraps for interval estimation.
(defaults to 1000). Larger numbers of bootstraps give a better
asymtotic CI, but may be very time-consuming (in particular for non-Gaussian data if some variance component 
is low). Bootstrapping can be switch off by setting \code{nboot = 0}.}

\item{npermut}{Number of permutations used when calculating asymptotic \emph{P} 
values (defaults to 0). Larger numbers of permutations give a better
asymtotic CI, but may be very time-consuming (in particular for non-Gaussian data if some variance component 
is low). Permutaton tests can be switch off by setting \code{npermut = 0}.}

\item{parallel}{If TRUE, bootstraps and permutations will be distributed across multiple cores.}

\item{ncores}{Specify number of cores to use for parallelization. On default,
all cores but one are used.}
}
\value{
Returns an object of class \code{rpt}. See details for specific functions.
}
\description{
A wrapper function for (adjusted) repeatability estimation from generalized linear mixed-effects 
models fitted by restricted maximum likelihood (REML). Calls 
specialised functions depending of the choice of datatype and method.
}
\details{
For \code{datatype='Gaussian'} calls function \link{rptGaussian},
         for \code{datatype='Poisson'} calls function \link{rptPoisson}, 
         for \code{datatype='Binary'} calls function \link{rptBinary}, 
         for \code{datatype='Proportion'} calls function \link{rptProportion}.
}
\examples{
# load data
data(BeetlesBody)
data(BeetlesMale)
data(BeetlesFemale)

#  prepare proportion data
BeetlesMale$Dark <- BeetlesMale$Colour
BeetlesMale$Reddish <- (BeetlesMale$Colour-1)*-1
md <- aggregate(cbind(Dark, Reddish) ~ Population + Container, data=BeetlesMale, FUN=sum)


# Note: all function calls are run without bootstraps and permutation for speed reasons

# gaussian data
rpt(BodyL ~ (1|Population), grname="Population", data=BeetlesBody, 
nboot=0, npermut=0, datatype = "Gaussian")

# poisson data
rpt(Egg ~ Treatment + (1|Container), grname=c("Container"), data = BeetlesFemale,
nboot=0, npermut=0, datatype = "Poisson")

# binary data
rpt(Colour ~ (1|Population), grname=c("Population"), 
data=BeetlesMale, nboot=5, npermut=0, datatype = "Binary")

# proportion data
rpt(cbind(Dark, Reddish) ~ (1|Population), grname=c("Population"), data=md,
nboot=0, npermut=0, datatype = "Proportion")


}
\author{
Holger Schielzeth  (holger.schielzeth@uni-jena.de), 
        Shinichi Nakagawa (s.nakagawa@unsw.edu.au),
        Martin Stoffel (martin.adam.stoffel@gmail.com)
}
\references{
Nakagawa, S. & Schielzeth, H. (2011) \emph{Repeatability for 
     Gaussian and non-Gaussian data: a practical guide for biologists}. 
     Biological Reviews 85: 935-956.
}
\seealso{
\link{rptR}
}
\keyword{models}

