% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptBinary.R
\name{rptBinary}
\alias{rptBinary}
\title{GLMM-based Repeatability Using REML}
\usage{
rptBinary(formula, grname, data, link = c("logit", "probit"), CI = 0.95,
  nboot = 1000, npermut = 1000, parallel = FALSE, ncores = NULL)
}
\arguments{
\item{formula}{Formula as used e.g. by \link{glmer}. The grouping factor(s) of
interest needs to be included as a random effect, e.g. '(1|groups)'.
Covariates and additional random effects can be included to estimate adjusted repeatabilities.}

\item{grname}{A character string or vector of character strings giving the
name(s) of the grouping factor(s), for which the repeatability should
be estimated. Spelling needs to match the random effect names as given in \code{formula}.}

\item{data}{A dataframe that contains the variables included in the \code{formula}
and \code{grname} arguments.}

\item{link}{Link function. \code{log} and \code{sqrt} are allowed, defaults to \code{log}.}

\item{CI}{Width of the confidence interval (defaults to 0.95).}

\item{nboot}{Number of parametric bootstraps for interval estimation.
Defaults to 1000. Larger numbers of permutations give a better
asymtotic CI, but may be very time-consuming.}

\item{npermut}{Number of permutations used when calculating 
asymptotic \emph{P} values (defaults to 1000).}

\item{parallel}{If TRUE, bootstraps will be distributed.}

\item{ncores}{Specify number of cores to use for parallelization. On default,
all cores but one are used.}
}
\value{
Returns an object of class rpt that is a a list with the following elements: 
\item{call}{function call}
\item{datatype}{Response distribution (here: 'Binomial')}.
\item{CI}{Width of the confidence interval.}
\item{R}{\code{data.frame} with point estimates for repeatabilities. Columns
     are groups of interest. Rows are original and link scale, in this order.}
\item{se}{\code{data.frame} with approximate standard errors (\emph{se}) for repeatabilities. Columns
     are groups of interest. Rows are original and link scale, in this order.
     Note that the distribution might not be symmetrical, in which case the \emph{se} is less informative.}
\item{CI_emp}{\code{list} of two elements containing the confidence intervals for repeatabilities 
     on the link and original scale, respectively. Within each list element, lower and upper CI
     are columns and each row is a grouping factor of interest.}
\item{P}{Approximate \emph{P} \code{data.frame} with p-values from a significance test based on likelihood-ratio
     in the first column and significance test based on permutation of residuals for 
     both the original and link scale in the second and third column. Each row is a grouping
     factors.}
\item{R_boot_link}{Parametric bootstrap samples for \emph{R} on the link scale. Each \code{list}
      element is a grouping factor.}
\item{R_boot_org}{Parametric bootstrap samples for \emph{R} on the original scale. Each \code{list}
      element is a grouping factor.}
\item{R_permut_link}{Permutation samples for \emph{R} on the link scale. Each \code{list}
      element is a grouping factor.}
\item{R_permut_org}{Permutation samples for \emph{R} on the original scale. Each \code{list}
      element is a grouping factor.}
\item{LRT}{List of Likelihood-ratios for the model and the reduced model(s), 
      and \emph{P} value(s) and degrees of freedom for the Likelihood-ratio test} 
\item{ngroups}{Number of groups.}
\item{nobs}{Number of observations.}
\item{overdisp}{Overdispersion parameter. Equals the variance in the observational factor random effect}
\item{mod}{Fitted model.}
}
\description{
Calculates repeatability from a general linear mixed-effects models fitted by REML (restricted maximum likelihood).
}
\examples{
 
# repeatability estimations for egg dumping (binary data)
data(BroodParasitism)
(rpt.BroodPar <- rptBinary(formula = cbpYN ~ (1|FemaleID), grname = c("FemaleID"), 
data = BroodParasitism))




nind = 80
nrep = 30 # a bit higher
latmu = 0
latbv = 0.3
latgv = 0.1
latrv = 0.2
indid = factor(rep(1:nind, each=nrep))
groid = factor(rep(1:nrep, nind))
latim = rep(rnorm(nind, 0, sqrt(latbv)), each=nrep)
latgm = rep(rnorm(nrep, 0, sqrt(latgv)), nind)
latvals = latmu + latim + latgm + rnorm(nind*nrep, 0, sqrt(latrv))
expvals = VGAM::logit(latvals, inverse = TRUE)
obsvals = rbinom(nind*nrep, 1, expvals)
beta0 = latmu
beta0 = VGAM::logit(mean(obsvals))
md = data.frame(obsvals, indid, groid)

R_est <- rptBinary(formula = obsvals ~ (1|indid) + (1|groid), grname = c("indid", "groid"), 
                    data = md, nboot = 3, link = "logit", npermut = 20, parallel = FALSE)
R_est2 <- rptBinary(formula = obsvals ~ (1|indid), grname = "indid", 
                    data = md, nboot = 10, link = "logit", npermut = 10, parallel = FALSE)
                    
}
\author{
Holger Schielzeth  (holger.schielzeth@ebc.uu.se),
        Shinichi Nakagawa (shinichi.nakagawa@otago.ac.nz) &
        Martin Stoffel (martin.adam.stoffel@gmail.com)
}
\references{
Carrasco, J. L. and Jover, L.  (2003). \emph{Estimating the generalized 
concordance correlation coefficient through variance components}. Biometrics 59: 849-858.

Faraway, J. J. (2006). \emph{Extending the linear model with R}. Boca Raton, FL, Chapman & Hall/CRC.

Nakagawa, S. and Schielzeth, H. (2010) \emph{Repeatability for Gaussian and 
non-Gaussian data: a practical guide for biologists}. Biological Reviews 85: 935-956
}
\seealso{
\link{rpt}
}

