% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptPoisson.R
\name{rptPoisson}
\alias{rptPoisson}
\title{GLMM-based Repeatability Using REML for Possoin-distributed data}
\usage{
rptPoisson(formula, grname, data, link = c("log", "sqrt"), CI = 0.95,
  nboot = 1000, npermut = 0, parallel = FALSE, ncores = NULL,
  ratio = TRUE)
}
\arguments{
\item{formula}{Formula as used e.g. by \link{lmer}. The grouping factor(s) of
interest needs to be included as a random effect, e.g. '(1|groups)'.
Covariates and additional random effects can be included to estimate adjusted 
repeatabilities.}

\item{grname}{A character string or vector of character strings giving the
name(s) of the grouping factor(s), for which the repeatability should
be estimated. Spelling needs to match the random effect names as given in \code{formula} 
and terms have to be set in quotation marks.}

\item{data}{A dataframe that contains the variables included in the \code{formula}
and \code{grname} arguments.}

\item{link}{Link function. \code{log} and \code{sqrt} are allowed, defaults to \code{log}.}

\item{CI}{Width of the required confidence interval (defaults to 0.95).}

\item{nboot}{Number of parametric bootstraps for interval estimation
(defaults to 1000). Larger numbers of bootstraps give a better
asymtotic CI, but may be very time-consuming (in particular of some variance component 
is low). Bootstrapping can be switch off by setting \code{nboot = 0}.}

\item{npermut}{Number of permutations used when calculating asymptotic p-values 
(defaults to 0). Larger numbers of permutations give a better
asymtotic CI, but may be very time-consuming (in particular of some variance component 
is low). Permutaton tests can be switch off by setting \code{npermut = 0}.}

\item{parallel}{If TRUE, bootstraps and permutations will be distributed across multiple cores.}

\item{ncores}{Specify number of cores to use for parallelization. On default,
all cores but one are used.}

\item{ratio}{Defaults to TRUE. If FALSE, the variance(s) of the grouping factor(s) of interest
will be used for all further calculations. The resulting point estimate(s), 
uncertainty interval(s) and significance test(s) therefore refer to the estimated variance
itself rather than to the repeatability (i.e. ratio of variances).}
}
\value{
Returns an object of class \code{rpt} that is a a list with the following elements: 
\item{call}{Function call}
\item{datatype}{Response distribution (here: 'Poisson').}
\item{CI}{Coverage of the confidence interval as specified by the \code{CI} argument.}
\item{R}{\code{data.frame} with point estimates for repeatabilities. Columns
     represent grouping factors of interest. Rows show original and link scale repeatabilites 
     (in this order).}
\item{se}{\code{data.frame} with approximate standard errors (\emph{se}) for repeatabilities. Columns
     are groups of interest. Rows are original and link scale (in this order).
     Note that the distribution might not be symmetrical, in which case the \emph{se} is less informative.}
\item{CI_emp}{\code{list} of two elements containing the confidence intervals for repeatabilities 
     on the link and original scale, respectively. Within each list element, lower and upper CI
     are columns and each row for each grouping factor of interest.}
\item{P}{\code{data.frame} with p-values from a significance test based on likelihood-ratios
     in the first column and significance test based on permutation of residuals for 
     both the original and link scale in the second and third column. Each row represents a grouping
     factor of interest.}
\item{R_boot_link}{Parametric bootstrap samples for \emph{R} on the link scale. Each \code{list}
      element is a grouping factor.}
\item{R_boot_org}{Parametric bootstrap samples for \emph{R} on the original scale. Each \code{list}
      element is a grouping factor.}
\item{R_permut_link}{Permutation samples for \emph{R} on the link scale. Each \code{list}
      element is a grouping factor.}
\item{R_permut_org}{Permutation samples for \emph{R} on the original scale. Each \code{list}
      element is a grouping factor.}
\item{LRT}{List of likelihoods for the full model and the reduced model(s), likelihood ratios \emph{D}, 
     p-value(s) and degrees of freedom for the likelihood-ratio test.} 
\item{ngroups}{Number of groups.}
\item{nobs}{Number of observations.}
\item{mod}{Fitted model.}
}
\description{
Estimates repeatability from a generalized linear mixed-effects models fitted by restricted maximum likelihood (REML).
}
\examples{
# load data
data(BeetlesFemale)

# Note: nboot and npermut are set to 5 for speed reasons. Use larger numbers
# for the real analysis.

rptPoisson(Egg ~ Treatment + (1|Container), grname=c("Container"), data = BeetlesFemale,
nboot=0, npermut=0)

rptPoisson(Egg ~ Treatment + (1|Container), grname=c("Container"), data = BeetlesFemale,
nboot=5, npermut=5, ratio = FALSE)

test <-  rptPoisson(formula = Egg ~ Treatment + (1|Container) + (1|Habitat) , grname=c("Container", "Habitat", "Residual", "Overdispersion"), data = BeetlesFemale,
nboot=5, npermut=5, ratio = TRUE)
}
\author{
Holger Schielzeth  (holger.schielzeth@uni-jena.de),
        Shinichi Nakagawa (s.nakagawa@unsw.edu.au) &
        Martin Stoffel (martin.adam.stoffel@gmail.com)
}
\references{
Carrasco, J. L. & Jover, L.  (2003) \emph{Estimating the generalized 
concordance correlation coefficient through variance components}. Biometrics 59: 849-858.

Faraway, J. J. (2006) \emph{Extending the linear model with R}. Boca Raton, FL, Chapman & Hall/CRC.

Nakagawa, S. & Schielzeth, H. (2010) \emph{Repeatability for Gaussian and 
non-Gaussian data: a practical guide for biologists}. Biological Reviews 85: 935-956
}
\seealso{
\link{rpt}
}

