\name{rptR}
\alias{rptR}
\title{
	Repeatability Calculation for Gaussian and Non-Gaussian Ddata
}
\description{
	A collection of functions for caluculating point estimates, interval estimates and significance tests of the repeatability (intra-class correlation coefficient) of measurements.
	The function \link{rpt} is a the core functions that calls more specialised functions as required. Specialised functions can also be called directly (see \link{rpt} for details).
	All functions return lists of values. The functions \link{print.rpt} produces summaries in a more appealing format.
}
\author{
	Holger Schielzeth  (holger.schielzeth@ebc.uu.se) & Shinichi Nakagawa (shinichi.nakagawa@otago.ac.nz)
}
\references{
	Nakagawa, S. and Schielzeth, H. (2010) \emph{Repeatability for Gaussian and non-Gaussian data: a practical guide for biologists}. Biological Reviews 85: 935-956
}
\keyword{package}